Add option to update role
Role can be deleted if it was created by accident. Role title can be
updated in case it changes or was inputted incorrectly to begin with.
Roles can also be marked as inactive if the team member is no longer
acting in this role. Inactive roles are not deleted, as they'll be used
to calculate the number of roles at different time periods.